+++ src/components/SwapForm.tsx
@@ -1,39 +1,6 @@
 import { useEffect, useState } from "react";
 import { Form, Input, message, Button } from "antd";
-import MakeSOLTrx from "../logic/transaction/MakeSOLTrx";
-import MakeNEARTrx from "../logic/transaction/MakeNEARTrx";
-import MakeTONTrx from "../logic/transaction/MakeTONTrx";
-import MakeATOMTrx from "../logic/transaction/MakeATOMTrx";
-import MakeUSNTrx from "../logic/transaction/MakeUSNTrx";
-import MakeAURORATrx from "../logic/transaction/MakeAURORATrx";
-import MakeETHTrx from "../logic/transaction/MakeETHTrx";
-
-import useResponsive from "../logic/Responsive";
-import MakeTONJettonsBurnTrx from "../logic/transaction/MakeTONJettonsBurnTrx";
-
-// au,
-// su,
-// tu,
-// nu,
-// ATOMwalletKey,
-// SOLwalletKey,
-// TONwalletKey,
-// NEARwalletKey,
-// ATOMMaxAmount,
-// SOLMaxAmount,
-// ETHMaxAmount,
-// TONMaxAmount,
-// NEARMaxAmount,
-// btnSelectSource,
-// btnSelectDirection,
-// btnDest,
-// btnSource,
-// setIsload,
-// isload,
-// hexString,
-// changeDirection,
-// directionNetwork
-// networkSource
+import makeTrx from "../logic/trxBuilder";
 
 const SwapForm = (props: any) => {
 	const [addVal, setAddVal] = useState("");
@@ -77,34 +44,6 @@ const SwapForm = (props: any) => {
 		isDirwATOMTON ||
 		isDirwUSNTON;
 	console.log(isTargetWrapp);
-
-	const sourceChain = isSouwSOLTON
-		? "TONwSOL"
-		: isSouwATOMTON
-		? "TONwATOM"
-		: isSouwNEARTON
-		? "TONwNEAR"
-		: isSouwAURTON
-		? "TONwAURORA"
-		: isSouwUSNTON
-		? "TONwUSN"
-		: isSouwETHTON
-		? "TONwETH"
-		: "";
-
-	const TONJettonContractAdd = isSouwSOLTON
-		? "EQC4cCygTZPKIP9cCsWx7DW5i5MQPOsEcfKkKwBZKkRCCfaW"
-		: isSouwATOMTON
-		? "EQCa5-xswEfQM5x_CBb5f53ghfy8ZYTAMCohgqSO6rBYMlkD"
-		: isSouwNEARTON
-		? "EQALr-K836vMmF5gOBzYmEHlS29-iG6AGsmHFzzgpMiy9ERi"
-		: isSouwAURTON
-		? "EQAlLZSs3HbZ6W5CoesPbqBoBLfS88FG1T0kLwaCC3fRF3ut"
-		: isSouwUSNTON
-		? "EQAfuJx-GWk0rn4T1r3g6SKmXRwBnW7I4jG2izu2qdoNH4aI"
-		: isSouwETHTON
-		? "EQB6l24gEV_OIR0IlZHpoWAnNzj-xS2Nf_uSAEcTx_7B4k_U"
-		: "";
 
 	useEffect(() => {
 		setParams("");
@@ -208,26 +147,6 @@ const SwapForm = (props: any) => {
 			: null
 	);
 
-	const TRXDir = (
-		props.directionNetwork === "sol"
-			? "SOLANA"
-			: props.directionNetwork === "wnear (ton)"
-			? "TONwNEAR"
-			: props.directionNetwork === "wsol (ton)"
-			? "TONwSOL"
-			: props.directionNetwork === "watom (ton)"
-			? "TONwATOM"
-			: props.directionNetwork === "waurora (ton)"
-			? "TONwAURORA"
-			: props.directionNetwork === "weth (ton)"
-			? "TONwETH"
-			: props.directionNetwork === "wusn (ton)"
-			? "TONwUSN"
-			: props.directionNetwork === "atom"
-			? "COSMOS"
-			: props.directionNetwork
-	).toUpperCase();
-
 	const sourceCurrencyName = isSouAtom
 		? "ATOM"
 		: isSouNear
@@ -292,90 +211,9 @@ const SwapForm = (props: any) => {
 		(openData ? !!params : true) &&
 		(openData ? !!addVal : true);
 
-	const TONTrx = () =>
-		MakeTONTrx(
-			activeBtn,
-			props.setIsload,
-			props.firstCurrAmount,
-			walletDirKey,
-			TRXDir,
-			props.hexString,
-			openData,
-			addVal,
-			params
-		);
-
 	useEffect(() => {
 		if (openData) setAddVal(walletDirKey);
 	}, [openData, walletDirKey]);
-
-	const SOLtrx = () =>
-		MakeSOLTrx(
-			activeBtn,
-			props.setIsload,
-			props.connection,
-			props.SOLwalletKey,
-			walletDirKey,
-			TRXDir,
-			props.firstCurrAmount
-		);
-
-	const NEARTrx = () =>
-		MakeNEARTrx(
-			activeBtn,
-			props.setIsload,
-			props.NEARwalletKey,
-			props.firstCurrAmount,
-			walletDirKey,
-			TRXDir,
-			props.hexString,
-			openData,
-			addVal,
-			params
-		);
-
-	const ATOMtrx = () =>
-		MakeATOMTrx(
-			activeBtn,
-			props.setIsload,
-			props.connection,
-			props.ATOMwalletKey,
-			walletDirKey,
-			TRXDir,
-			props.firstCurrAmount
-		);
-
-	const USNtrx = () =>
-		MakeUSNTrx(
-			activeBtn,
-			props.setIsload,
-			props.NEARwalletKey,
-			props.firstCurrAmount,
-			walletDirKey,
-			TRXDir,
-			props.hexString,
-			openData,
-			addVal,
-			params
-		);
-
-	const AURORAtrx = () =>
-		MakeAURORATrx(props.setIsload, walletDirKey, TRXDir, props.firstCurrAmount);
-
-	const ETHtrx = () =>
-		MakeETHTrx(props.setIsload, walletDirKey, TRXDir, props.firstCurrAmount);
-
-	const TONJettonsBurnTrx = () =>
-		MakeTONJettonsBurnTrx(
-			sourceChain,
-			TONJettonContractAdd,
-			activeBtn,
-			props.setIsload,
-			props.firstCurrAmount,
-			props.TONwalletKey,
-			TRXDir,
-			walletDirKey
-		);
 
 	useEffect(() => {
 		props.setFirstCurrAmount("");
@@ -417,8 +255,6 @@ const SwapForm = (props: any) => {
 					}
 					placeholder={"0.000"}
 				/>
-				{/* {!isMobile ? props.btnSelectSource : null}
-				{!isMobile ? props.btnSource : null} */}
 			</Form.Item>
 			{props.changeDirection}
 			<Form.Item label={`TO`}>
@@ -447,8 +283,6 @@ const SwapForm = (props: any) => {
 					}}
 					placeholder={"0.000"}
 				/>
-				{/* {!isMobile ? props.btnSelectDirection : null}
-				{!isMobile ? props.btnDest : null} */}
 			</Form.Item>
 			{addMessage ? (
 				<>
@@ -496,43 +330,35 @@ const SwapForm = (props: any) => {
 			{((currency / secCurrency) * 0.975).toFixed(3)} {directionCurrencyName}
 			<br />
 			Tonana reserve: {MaxDirAmount.toFixed(3)} {directionCurrencyName}
-			{/* <br />
-			You will get{" "}
-			{!!Number(otherAmount)
-				? (Number(otherAmount) * 0.975).toFixed(6)
-				: "0.000000"}{" "}
-			{direction} */}
 			<Form.Item
 				style={{
 					margin: "24px 0 0 0",
-					// filter: !activeBtn ? "grayscale(50%) contrast(50%)" : "",
 				}}>
 				<Button
 					type="primary"
 					id={activeBtn ? "submitBtn" : "nonactivesubmitBtn"}
-					onClick={
-						isSouAtom
-							? ATOMtrx
-							: isSouNear
-							? NEARTrx
-							: isSouUsn
-							? USNtrx
-							: isSouTon
-							? TONTrx
-							: isSouSol
-							? SOLtrx
-							: isSouAur
-							? AURORAtrx
-							: isSouEth
-							? ETHtrx
-							: isSouwNEARTON ||
-							  isSouwSOLTON ||
-							  isSouwATOMTON ||
-							  isSouwAURTON ||
-							  isSouwETHTON ||
-							  isSouwUSNTON
-							? TONJettonsBurnTrx
-							: () => {}
+					onClick={() =>
+						makeTrx(
+							activeBtn,
+							props,
+							walletDirKey,
+							openData,
+							addVal,
+							params,
+							isSouAtom,
+							isSouNear,
+							isSouUsn,
+							isSouTon,
+							isSouSol,
+							isSouAur,
+							isSouEth,
+							isSouwNEARTON,
+							isSouwSOLTON,
+							isSouwATOMTON,
+							isSouwAURTON,
+							isSouwETHTON,
+							isSouwUSNTON
+						)
 					}>
 					Submit
 				</Button>
